refactor: merge packages/bot into packages/api, rename to server
Flatten the two-package structure (bot + api) into a single server
package, eliminating the circular dependency workarounds (broadcast.ts,
client.ts) that existed solely to bridge the package boundary.

Changes:
- Rename packages/api -> packages/server
- Delete packages/bot (merged into server)
- Move bot/src/ contents into server/src/ (flattened, no sub-package split)
- Delete lib/broadcast.ts and lib/client.ts (indirection inlined)
- Update all @alfira-bot/bot imports to local module references
- Update Dockerfile, docker-compose, workflows, biome.json, docs
