+++ CLAUDE.md
@@ -4,14 +4,13 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 
 ## Project Overview
 
-Alfira is a self-hosted Discord music bot with a web UI as the primary interface. It's a Bun workspaces monorepo with four packages:
+Alfira is a self-hosted Discord music bot with a web UI as the primary interface. It's a Bun workspaces monorepo with three packages:
 
 - `packages/shared` — Shared types, utilities, DB schema, and logger
-- `packages/bot` — Discord bot (`GuildPlayer`, NodeLink audio via `hoshimi`)
-- `packages/api` — Bun API server (HTTP + WebSocket, Drizzle ORM)
+- `packages/server` — Bun API server + Discord bot (`GuildPlayer`, NodeLink audio via `hoshimi`, Seyfert v4)
 - `packages/web` — React 19 + Tailwind CSS 4 web UI
 
-The bot and API run in a **single Bun process** started from `packages/api/src/index.ts`. They share memory for player state, enabling real-time WebSocket broadcasts directly from playback events.
+The bot and API run in a **single Bun process** started from `packages/server/src/index.ts`. They share memory for player state, enabling real-time WebSocket broadcasts directly from playback events.
 
 ## Tech Stack
 
@@ -27,7 +26,7 @@ The bot and API run in a **single Bun process** started from `packages/api/src/i
 ## Development Commands
 
 ```bash
-# Build shared + bot dist/, then start all services with Docker
+# Build shared + server dist/, then start all services with Docker
 bun run dev
 
 # Build the web UI (needed after web/src changes before docker compose restart)
@@ -51,27 +50,27 @@ bun run format
 
 ## Key Architecture Notes
 
-### Single-Process Startup Sequence (packages/api/src/index.ts)
+### Single-Process Startup Sequence (packages/server/src/index.ts)
 
 1. Run database migrations (homegrown, reads `packages/shared/dist/db/migrations/*.sql`)
 2. Verify database connectivity
-3. Call `setBroadcastQueueUpdate(emitPlayerUpdate)` — injects the WebSocket broadcaster into the bot package
-4. Call `startBot()` — initializes Seyfert Discord client + Hoshimi + NodeLink connection
+3. Start NodeLink subprocess
+4. Call `startDiscord()` — initializes Seyfert Discord client + Hoshimi + NodeLink connection
 5. Start Bun HTTP server on port 3001 (serves API routes, WebSocket at `/ws`, and static web assets from `packages/web/dist/`)
 
 ### Real-Time Updates (WebSocket Pipeline)
 
 The bot never directly holds WebSocket connections. Instead:
 
-1. `GuildPlayer` calls `broadcastQueueUpdate(state)` (packages/bot/src/lib/broadcast.ts)
-2. `broadcastQueueUpdate` calls the injected `emitPlayerUpdate(state)` function
-3. `emitPlayerUpdate` (packages/api/src/lib/socket.ts) sends to all registered WebSocket clients
+1. `GuildPlayer` calls `broadcastQueueUpdate(state)` (packages/server/src/startDiscord.ts)
+2. `broadcastQueueUpdate` calls `emitPlayerUpdate(state)` directly (same package)
+3. `emitPlayerUpdate` (packages/server/src/lib/socket.ts) sends to all registered WebSocket clients
 
 WebSocket clients authenticate via session cookie on connection. The client never sends messages — it's receive-only.
 
-### Bot Changes Require Rebuild
+### Server Changes Require Rebuild
 
-The bot is compiled to `packages/bot/dist/` during Docker image build. If you change `packages/bot/src/**`, run `bun run dev` again — it rebuilds the local `dist/` and then starts Docker with a fresh image. API source is live-mounted via Docker volume so `docker compose restart alfira` picks up changes without a rebuild.
+The server is compiled to `packages/server/dist/` during Docker image build. If you change `packages/server/src/**`, run `bun run dev` again — it rebuilds the local `dist/` and then starts Docker with a fresh image. API source is live-mounted via Docker volume so `docker compose restart alfira` picks up changes without a rebuild.
 
 ### Environment Configuration
 
@@ -79,7 +78,7 @@ A single `.env` file at the project root is used for all configuration. Copy `.e
 
 ### NodeLink Audio Service
 
-The bot streams audio from NodeLink (a Lavalink v4-compatible server). The `nodelink` service runs in Docker (docker-compose.yml) on port 2333. `hoshimi` (packages/bot) manages players and voice connections. When editing audio-related code, be aware that `stop(true)` on a Hoshimi player destroys the voice session on NodeLink, which can cause the "No voice state, track is enqueued" issue if the next track tries to play before reconnecting.
+The bot streams audio from NodeLink (a Lavalink v4-compatible server). The `nodelink` service runs in Docker (docker-compose.yml) on port 2333. `hoshimi` (packages/server) manages players and voice connections. When editing audio-related code, be aware that `stop(true)` on a Hoshimi player destroys the voice session on NodeLink, which can cause the "No voice state, track is enqueued" issue if the next track tries to play before reconnecting.
 
 ## Code Style
