--- packages/bot/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@alfira-bot/bot",
-  "version": "0.1.0",
-  "private": true,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": "./dist/index.js"
-  },
-  "scripts": {
-    "deploy": "bun run src/deploy-commands.ts",
-    "build": "tsc"
-  },
-  "dependencies": {
-    "@alfira-bot/shared": "workspace:*",
-    "dotenv": "^17.3.1",
-    "hoshimi": "^0.3.9",
-    "seyfert": "^4.3.0"
-  },
-  "devDependencies": {
-    "@types/node": "^25.4.0",
-    "typescript": "^6.0.0"
-  }
-}
