--- packages/bot/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "ignoreDeprecations": "6.0",
-    "baseUrl": ".",
-    "rootDir": "./src",
-    "paths": {
-      "@alfira-bot/shared": ["./node_modules/@alfira-bot/shared/dist/"],
-      "@alfira-bot/shared/*": ["./node_modules/@alfira-bot/shared/dist/*"]
-    },
-    "lib": ["ES2020"],
-    "outDir": "./dist",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "declaration": true,
-    "declarationMap": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
