+++ packages/server/src/lib/serialization.ts
