+++ packages/server/src/middleware/requireAuth.ts
