+++ packages/server/src/routes/auth.ts
